exclude long-job-names warning for checkconfig
